requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosmtpd==1.4.6
annotated-types==0.7.0
anyio==4.9.0
atpublic==5.1
attrs==25.3.0
base58==2.1.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coincurve==21.0.0
cryptography==44.0.2
dkimpy==1.1.8
dnspython==2.7.0
-e git+https://github.com/ManticoreTechnologies/evrmail.git@3e03a3acd90118325675d8ea0e67b2bad7546408#egg=evrmail
evrmore-rpc==4.0.0
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
idna==3.10
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
multidict==6.2.0
platformdirs==4.3.7
propcache==0.3.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.1
textual==3.0.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
